refactor: 멤버 서비스 테스트와 멤버 컨트롤러 테스트 병합
